refactor(createBookInfo): tidy up implementation
