+++ src/createBookInfo/index.js
@@ -1,6 +1,6 @@
-const { getHTML } = require('./crawler') 
-const { getBookInfo } = require('./weread-parser')
-const { createReadNote } = require('./notion-integration') 
+const { getHTML } = require('../crawler') 
+const { getBookInfo } = require('../weread-parser')
+const { createReadNote } = require('../notion-integration') 
 
 async function createBookInfo() {
   const rawHtml = await getHTML()
