+++ src/createBookInfo/wereadParser.js
